feat/#7 : 폴더 구조 변경
